requirements.txt
asgiref==3.3.1
boto3==1.17.33
botocore==1.20.33
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.6
ddquery==0.1
defusedxml==0.7.1
Django==3.1.7
django-braces==1.14.0
django-cleanup==5.1.0
django-debug-toolbar==3.2
django-environ==0.4.5
django-extensions==3.1.1
django-filter==2.4.0
django-oauth-toolkit==1.4.1
django-request-logging==0.7.2
django-rest-framework-social-oauth2==1.1.0
django-storages==1.11.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
drf-access-policy==0.8.7
drf-nested-routers==0.93.3
drf-social-oauth2==1.0.9
drf-writable-nested==0.6.2
ecdsa==0.14.1
generics==3.3.0
idna==2.10
jmespath==0.10.0
oauthlib==3.1.0
Pillow==8.1.2
psycopg2-binary==2.8.6
pyasn1==0.4.8
pycparser==2.20
Pygments==2.2.0
PyJWT==2.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-jose==3.2.0
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
s3transfer==0.3.6
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.2.3
urllib3==1.26.4
